freiends_project/momentum-etf-main/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
apscheduler==3.11.0
    # via -r requirements.in
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.13.5
    # via
    #   -r requirements.in
    #   yfinance
certifi==2025.8.3
    # via
    #   curl-cffi
    #   requests
cffi==2.0.0
    # via curl-cffi
charset-normalizer==3.4.3
    # via requests
contourpy==1.3.2
    # via matplotlib
curl-cffi==0.13.0
    # via yfinance
cycler==0.12.1
    # via matplotlib
datetime==5.5
    # via pykrx
deprecated==1.2.18
    # via pykrx
dnspython==2.8.0
    # via pymongo
exchange-calendars==4.11.1
    # via pandas-market-calendars
fonttools==4.59.2
    # via matplotlib
frozendict==2.4.6
    # via yfinance
gitdb==4.0.12
    # via gitpython
idna==3.10
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
korean-lunar-calendar==0.3.1
    # via exchange-calendars
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.6
    # via pykrx
multipledispatch==1.0.0
    # via pykrx
multitasking==0.0.12
    # via yfinance
narwhals==2.5.0
    # via altair
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   exchange-calendars
    #   matplotlib
    #   pandas
    #   pykrx
    #   yfinance
packaging==24.2
    # via
    #   matplotlib
pandas==2.3.2
    # via
    #   -r requirements.in
    #   exchange-calendars
    #   pandas-market-calendars
    #   pykrx
    #   yfinance
# pandas-market-calendars==5.1.1  # 거래일 캘린더 생성은 수동 작업으로만 사용
    # via -r requirements.in
peewee==3.18.2
    # via yfinance
pillow==10.4.0
    # via
    #   matplotlib
platformdirs==4.4.0
    # via yfinance
protobuf==5.29.5
    # via yfinance
pyarrow==21.0.0
pycparser==2.23
    # via cffi
pykrx==1.2.4
    # via -r requirements.in
pyluach==2.3.0
    # via exchange-calendars
pymongo==4.15.1
    # via -r requirements.in
pyparsing==3.2.4
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
    #   pandas-market-calendars
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via
    #   datetime
    #   pandas
    #   yfinance
pyyaml==6.0.3
    # via exchange-calendars
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   pykrx
    #   yfinance
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
slack-sdk==3.36.0
    # via -r requirements.in
smmap==5.0.2
    # via gitdb
soupsieve==2.8
    # via beautifulsoup4
toolz==1.0.0
    # via exchange-calendars
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   referencing
tzdata==2025.2
    # via
    #   exchange-calendars
    #   pandas
    #   pandas-market-calendars
tzlocal==5.3.1
    # via apscheduler
urllib3==2.5.0
    # via requests
websockets==15.0.1
    # via yfinance
wrapt==1.17.3
    # via deprecated
xlrd==2.0.2
    # via pykrx
yfinance==0.2.65
    # via -r requirements.in
zope-interface==8.0
    # via datetime

# The following packages are considered to be unsafe in a requirements file:
# setuptools
